app/src/androidTest/java/MainFragmentTest.java
package com.example.vu.android.empowerplant;

import android.content.Context;
import android.os.Bundle;
import androidx.lifecycle.Lifecycle;
import io.sentry.ITransaction;
import io.sentry.Sentry;
import org.junit.Before;
import org.junit.Test;
import org.mockito.MockitoAnnotations;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;

import java.util.HashMap;

public class MainFragmentTest {

    private MainFragment mainFragment;

    @Before
    public void setUp() {
        MockitoAnnotations.initMocks(this);
        mainFragment = MainFragment.newInstance();
    }

    @Test
    public void testMainFragmentNotNull() {
        assertNotNull(mainFragment);
    }

    @Test
    public void testOnCreate() {
        mainFragment.onCreate(null);
        assertNotNull(mainFragment.getArguments());
    }

    @Test
    public void testOnCreateView() {
        assertNotNull(mainFragment.onCreateView(null, null, null));
    }

    @Test
    public void testSetBadgeNumber() {
        mainFragment.setBadgeNumber();
        // Assuming setBadgeNumber increases the badge number by 1
        assertEquals(1, mainFragment.mCartItemCount);
    }

    @Test
    public void testFetchToolsFromServer() {
        mainFragment.fetchToolsFromServer();
        // Assuming fetchToolsFromServer fetches tools, we can check if the progress dialog is shown
        assertNotNull(mainFragment.progressDialog);
        assertTrue(mainFragment.progressDialog.isShowing());
    }

    @Test
    public void testOnItemClick() {
        StoreItem storeItem = new StoreItem();
        storeItem.setQuantity(1);
        mainFragment.onItemClick(storeItem);
        // Assuming onItemClick increases the item count
        assertEquals(1, storeItem.getQuantity());
    }

    @Test
    public void testProcessGetToolsResponse() {
        mainFragment.processGetToolsResponse("");
        // Assuming processGetToolsResponse updates the adapter
        assertNotNull(mainFragment.adapter);
    }

    @Test
    public void testProcessProducts() {
        mainFragment.processProducts();
        // Assuming processProducts calls getIterator, which returns a value
        assertEquals(1, mainFragment.getIterator(1));
    }

    @Test
    public void testGetIterator() {
        assertEquals(1, mainFragment.getIterator(1));
    }

    @Test
    public void testGetEmpowerPlantDomain() {
        assertNotNull(mainFragment.getEmpowerPlantDomain());
    }

    @Test
    public void testCheckout() {
        mainFragment.checkout();
        // Assuming checkout starts a transaction, check if the transaction is not null
        ITransaction transaction = Sentry.startTransaction("checkout [android]", "http.client");
        assertNotNull(transaction);
    }

    @Test
    public void testBuildJSONPostData() {
        HashMap<String, StoreItem> storeItems = new HashMap<>();
        assertNotNull(mainFragment.buildJSONPostData(storeItems));
    }

    @Test
    public void testProcessDeliveryItem() {
        ITransaction transaction = Sentry.startTransaction("test", "test");
        mainFragment.processDeliveryItem(transaction);
        // Assuming processDeliveryItem sets the status
        assertEquals(transaction.getStatus(), SpanStatus.OK);
    }

    @Test
    public void testAddAttachment() {
        assertTrue(mainFragment.addAttachment());
    }

    @Test
    public void testInsertMultipleStoreItems() {
        mainFragment.insertMultipleStoreItems();
        // Assuming insertMultipleStoreItems interacts with a database, this is harder to assert directly
    }
}
